# weathra/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AllowanceListResponse": {
        "additionalProperties": false,
        "properties": {
          "allowances": {
            "items": {
              "$ref": "#/components/schemas/AllowanceRecord"
            },
            "title": "Allowances",
            "type": "array"
          },
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count",
          "allowances"
        ],
        "title": "AllowanceListResponse",
        "type": "object"
      },
      "AllowanceRecord": {
        "additionalProperties": false,
        "description": "One allowance: a subject, a dimension, a window, a number.",
        "properties": {
          "allowance": {
            "minimum": 0.0,
            "title": "Allowance",
            "type": "integer"
          },
          "dimension": {
            "$ref": "#/components/schemas/QuotaDimension"
          },
          "internal_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Subject"
          },
          "plan_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "window_kind": {
            "$ref": "#/components/schemas/QuotaWindow"
          }
        },
        "required": [
          "dimension",
          "window_kind",
          "allowance"
        ],
        "title": "AllowanceRecord",
        "type": "object"
      },
      "AllowanceRequest": {
        "additionalProperties": false,
        "description": "One allowance, for a plan or for the internal subject.\n\nThe value is bounded above as well as below. A negative allowance is refused because\n`specs/usage-limits` says so; an absurd one is refused because a typo that adds three zeroes to\na monthly token budget is a bill, and there is no legitimate allowance near the ceiling.",
        "properties": {
          "allowance": {
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Allowance",
            "type": "integer"
          },
          "dimension": {
            "$ref": "#/components/schemas/QuotaDimension"
          }
        },
        "required": [
          "dimension",
          "allowance"
        ],
        "title": "AllowanceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanAssignmentRequest": {
        "additionalProperties": false,
        "description": "Which tier a person is on.\n\n``PlanCode`` and not a string, so a body naming ``plus`` is refused by validation before the\nhandler runs. The retired tier is not a value this enum has and no alias admits it.",
        "properties": {
          "plan_code": {
            "$ref": "#/components/schemas/PlanCode"
          }
        },
        "required": [
          "plan_code"
        ],
        "title": "PlanAssignmentRequest",
        "type": "object"
      },
      "PlanCode": {
        "description": "The canonical subscription plan codes.\n\nThree tiers, lowercase, stable enough to map to an external billing product later\n(``specs/usage-limits``' \"room left for a later billing integration\"). The display names live\nin ``subscription_plans`` as data, so renaming *Premium* on a pricing page is a row update; the\ncode here is the identifier that must not move, because `user_plans` rows and recorded usage\nevents both carry it.\n\nThere is deliberately **no `plus`**. An earlier draft of the specs named the middle tier that\nway; the product decision of 2026-09-09 settled on Free / Pro / Premium and the specs were\nreconciled to match. No alias is provided in either direction \u2014 this foundation has never\nshipped, so there is no customer contract to be backward-compatible with, and an alias would\nonly make two names for one tier that a later reader has to reconcile again.",
        "enum": [
          "free",
          "pro",
          "premium"
        ],
        "title": "PlanCode",
        "type": "string"
      },
      "PlanListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanRecord"
            },
            "title": "Plans",
            "type": "array"
          }
        },
        "required": [
          "count",
          "plans"
        ],
        "title": "PlanListResponse",
        "type": "object"
      },
...
      "PlanRecord": {
        "additionalProperties": false,
        "description": "A product tier, its rank, and the policy it maps each call role to.",
        "properties": {
          "display_name": {
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "external_subscription_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unused. Where a billing provider's id would later land.",
            "title": "External Subscription Ref"
          },
          "plan_code": {
            "$ref": "#/components/schemas/PlanCode"
          },
          "policy_by_call_role": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/CallRole"
            },
            "title": "Policy By Call Role",
            "type": "object"
          },
          "rank": {
            "description": "Ascending entitlement. What 'never escalate above' compares.",
            "minimum": 0.0,
            "title": "Rank",
            "type": "integer"
          }
        },
        "required": [
          "plan_code",
          "display_name",
          "rank"
        ],
        "title": "PlanRecord",
        "type": "object"
      },
...
      "QuotaWindow": {
        "description": "The period an allowance is counted over.\n\n``CONCURRENT`` is not a calendar window and is named so it cannot be mistaken for one: it is a\nlive gauge incremented on run start and decremented in a ``finally``, with no reset boundary\nbecause it has nothing to reset.",
        "enum": [
          "day",
          "month",
          "concurrent"
        ],
        "title": "QuotaWindow",
        "type": "string"
      },
...
      "RoleGrantResponse": {
        "additionalProperties": false,
        "description": "One role grant, as an administrator sees it. No contact detail, because none is stored.",
        "properties": {
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "granted_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null for the bootstrap grant, which has no granter to name.",
            "title": "Granted By"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "subject_id",
          "role",
          "granted_at"
        ],
        "title": "RoleGrantResponse",
        "type": "object"
      },
      "RoleListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "grants": {
            "items": {
              "$ref": "#/components/schemas/RoleGrantResponse"
            },
            "title": "Grants",
            "type": "array"
          }
        },
        "required": [
          "count",
          "grants"
        ],
        "title": "RoleListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SupabaseAccessToken": {
        "bearerFormat": "JWT",
        "description": "A Supabase Auth access token, presented as `Authorization: Bearer <token>`. Weathra validates the signature against the project's published signing keys and never issues, stores, or refreshes a credential of its own.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Agentic weather intelligence. Every figure in a Weathra answer is retrieved from a named provider\nor computed by deterministic code, and every answer carries the evidence record to prove it.\n\nEndpoints are classified **public** or **protected**. A public endpoint takes parameters and\nreturns weather; it reads no user-owned data, except that it applies a signed-in caller's unit\npreference. A protected endpoint requires a Supabase access token as a Bearer credential and\noperates on the acting user's own data.\n",
    "title": "Weathra",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/allowances": {
      "get": {
        "description": "**Protected. Administrative.** Lists the usage allowances, per plan and for the internal subject.\n\nEvery allowance, or one plan's, or the internal subject's.\n\nA dimension with no row is *unlimited*, and it is absent here rather than reported as a large\nnumber \u2014 filling it in would be reporting a limit nobody configured.",
        "operationId": "list_allowances_api_v1_admin_allowances_get",
        "parameters": [
          {
            "description": "Narrow to one plan.",
            "in": "query",
            "name": "plan_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlanCode"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow to one plan.",
              "title": "Plan Code"
            }
          },
          {
            "description": "The internal allowance only.",
            "in": "query",
            "name": "internal",
            "required": false,
            "schema": {
              "default": false,
              "description": "The internal allowance only.",
              "title": "Internal",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllowanceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "List the allowances",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/allowances/internal": {
      "put": {
        "description": "**Protected. Administrative.** Sets one of the internal allowances that lab, evaluation and administrative traffic is accounted against.\n\nThe allowance administrative, lab and evaluation traffic is accounted against.\n\nIts own route rather than a flag on the one above, because the internal subject is not a plan\nand a body field that switched between them would be one typo away from raising Free's\nallowance while meaning to raise the lab's.",
        "operationId": "set_internal_allowance_api_v1_admin_allowances_internal_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllowanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Set one of the internal allowances",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/admin/plans": {
      "get": {
        "description": "**Protected. Administrative.** Lists the subscription plans.",
        "operationId": "list_plans_api_v1_admin_plans_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "List the subscription plans",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/plans/{plan_code}/allowances": {
      "put": {
        "description": "**Protected. Administrative.** Sets one of a plan's usage allowances.\n\nChange one allowance. Takes effect for the remainder of the current window.\n\nNot cached anywhere: `entitlements/quotas.py` reads `usage_limits` fresh on every admission,\nprecisely so \"an allowance change takes effect without a code change\" means *now* rather than\nwithin a staleness window.",
        "operationId": "set_plan_allowance_api_v1_admin_plans__plan_code__allowances_put",
        "parameters": [
          {
            "description": "free, pro, or premium.",
            "in": "path",
            "name": "plan_code",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlanCode",
              "description": "free, pro, or premium."
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllowanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Set one of a plan's allowances",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/admin/principals/administrators": {
      "get": {
        "description": "**Protected. Administrative.** Lists who holds the administrative role and who granted it.\n\nWho holds the administrative role, and who granted it to them.\n\nPrivileged, and only reachable here: the owner policy on `admin_roles` returns a request\nsession exactly its own row, so this list cannot be assembled from the request path however\nthe query is written.",
        "operationId": "list_administrators_api_v1_admin_principals_administrators_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "List the administrative principals",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/principals/{subject_id}/plan": {
      "put": {
        "description": "**Protected. Administrative.** Assigns a principal to a subscription plan.\n\nPut a person on a tier.\n\nThe only way an assignment happens: `user_plans` grants the request role `SELECT` and nothing\nelse, so a caller cannot assign themselves one however the request is shaped.",
        "operationId": "assign_plan_api_v1_admin_principals__subject_id__plan_put",
        "parameters": [
          {
            "description": "An auth subject.",
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "description": "An auth subject.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Subject Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Assign a principal to a plan",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/principals/{subject_id}/role": {
      "delete": {
        "description": "**Protected. Administrative.** Grants and revokes the administrative role.\n\nDemote a principal \u2014 but never yourself.\n\nThe self-revocation refusal is not paternalism about a mistake. An estate whose last\nadministrator demotes themselves has no way back in except the bootstrap script and a database\ncredential, and the person most likely to do it by accident is the one testing whether the\nendpoint works.",
        "operationId": "revoke_role_api_v1_admin_principals__subject_id__role_delete",
        "parameters": [
          {
            "description": "An auth subject.",
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "description": "An auth subject.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Subject Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Revoke a principal's administrative role",
        "tags": [
          "administration"
        ]
      },
      "put": {
        "description": "**Protected. Administrative.** Grants and revokes the administrative role.\n\nPromote a principal. Privileged, recorded, and idempotent.\n\n`specs/authentication` requires an administrative action to be attributed, and this is the\naction that most needs it: it is the one that creates the ability to perform every other one.",
        "operationId": "grant_role_api_v1_admin_principals__subject_id__role_put",
        "parameters": [
          {
            "description": "An auth subject.",
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "description": "An auth subject.",
              "maxLength": 64,
              "minLength": 1,
              "title": "Subject Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No valid access token was presented."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SupabaseAccessToken": []
          }
        ],
        "summary": "Grant a principal the administrative role",
        "tags": [
          "administration"
        ]
      }
    },
...
  }
}